refactor: UI 모델 패키지 이동
