+++ app/src/main/java/woowacourse/payments/view/cardaddition/CardAdditionActivity.kt
@@ -8,8 +8,8 @@ import androidx.activity.enableEdgeToEdge
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.ui.Modifier
 import woowacourse.payments.R
-import woowacourse.payments.view.CardUiModel
 import woowacourse.payments.view.cardaddition.component.CardAdditionScreen
+import woowacourse.payments.view.ui.model.CardUiModel
 import woowacourse.payments.view.ui.theme.AndroidpaymentsTheme
 
 class CardAdditionActivity : ComponentActivity() {
