+++ app/src/main/java/woowacourse/payments/view/cardaddition/CardAdditionUiState.kt
@@ -2,7 +2,7 @@ package woowacourse.payments.view.cardaddition
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
-import woowacourse.payments.view.CardUiModel
+import woowacourse.payments.view.ui.model.CardUiModel
 
 @Parcelize
 data class CardAdditionUiState(
