+++ app/src/main/java/woowacourse/payments/view/cardediting/CardEditingUiState.kt
@@ -3,7 +3,7 @@ package woowacourse.payments.view.cardediting
 import android.os.Parcelable
 import kotlinx.parcelize.IgnoredOnParcel
 import kotlinx.parcelize.Parcelize
-import woowacourse.payments.view.CardUiModel
+import woowacourse.payments.view.ui.model.CardUiModel
 
 @Parcelize
 data class CardEditingUiState(
