+++ app/src/main/java/woowacourse/payments/view/cardediting/component/CardEditingScreen.kt
@@ -16,8 +16,6 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import androidx.compose.ui.unit.dp
-import woowacourse.payments.view.BankTypeUiModel
-import woowacourse.payments.view.CardUiModel
 import woowacourse.payments.view.cardediting.CardEditingStateHolder
 import woowacourse.payments.view.cardediting.CardEditingUiEvent
 import woowacourse.payments.view.cardediting.CardEditingUiState
@@ -27,6 +25,8 @@ import woowacourse.payments.view.ui.component.CardOwnerNameTextField
 import woowacourse.payments.view.ui.component.ExpiredDateTextField
 import woowacourse.payments.view.ui.component.PasswordTextField
 import woowacourse.payments.view.ui.component.PaymentCard
+import woowacourse.payments.view.ui.model.BankTypeUiModel
+import woowacourse.payments.view.ui.model.CardUiModel
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
