+++ app/src/main/java/woowacourse/payments/view/cards/CardsStateHolder.kt
@@ -9,9 +9,9 @@ import androidx.compose.runtime.setValue
 import woowacourse.payments.domain.Card
 import woowacourse.payments.domain.CardsRepository
 import woowacourse.payments.domain.MockCardsRepository
-import woowacourse.payments.view.CardUiModel
 import woowacourse.payments.view.toDomain
 import woowacourse.payments.view.toUiModel
+import woowacourse.payments.view.ui.model.CardUiModel
 
 class CardsStateHolder(
     initialState: CardsUiState = CardsUiState(),
