+++ app/src/main/java/woowacourse/payments/view/cards/component/CardsScreen.kt
@@ -23,8 +23,6 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import woowacourse.payments.R
-import woowacourse.payments.view.BankTypeUiModel
-import woowacourse.payments.view.CardUiModel
 import woowacourse.payments.view.EXTRA_CARD
 import woowacourse.payments.view.cardaddition.CardAdditionActivity
 import woowacourse.payments.view.cardediting.CardEditingActivity
@@ -32,6 +30,8 @@ import woowacourse.payments.view.cards.CardsStateHolder
 import woowacourse.payments.view.cards.CardsUiEvent
 import woowacourse.payments.view.cards.CardsUiState
 import woowacourse.payments.view.cards.rememberCardsStateHolder
+import woowacourse.payments.view.ui.model.BankTypeUiModel
+import woowacourse.payments.view.ui.model.CardUiModel
 import woowacourse.payments.view.ui.theme.AndroidpaymentsTheme
 
 @Composable
