+++ app/src/main/java/woowacourse/payments/view/ui/model/CardUiModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.view
+package woowacourse.payments.view.ui.model
 
 import YearMonthParser
 import android.os.Parcelable
